ADD: CommunityController.java 작성 완료
