+++ social/src/main/java/daehee/challengehub/controller/CommunityInteractionController.java
@@ -138,131 +138,7 @@ public class CommunityInteractionController {
     }
 
 
-    // TODO: FriendRequestDto 구현해야할 듯, Friend 관련 컨트롤러도 분리가 필요해보임
-    // 친구 요청 목록 조회
-//    @GetMapping("/friends/requests")
-//    public ResponseEntity<List<FriendRequestDto>> getFriendRequests() {
-//        // 임의의 친구 요청 데이터 생성
-//        List<FriendRequestDto> friendRequests = Arrays.asList(
-//                FriendRequestDto.builder()
-//                        .requesterId(101L) // 요청자 ID
-//                        .requestedId(202L) // 요청 받은 사용자 ID
-//                        .requestDate("2023-11-10") // 요청 날짜
-//                        .status("대기 중") // 요청 상태
-//                        .build(),
-//                FriendRequestDto.builder()
-//                        .requesterId(103L) // 다른 요청자 ID
-//                        .requestedId(202L) // 같은 요청 받은 사용자 ID
-//                        .requestDate("2023-11-09") // 요청 날짜
-//                        .status("대기 중") // 요청 상태
-//                        .build()
-//        );
-//        return ResponseEntity.ok(friendRequests);
-//    }
-
-    // TODO: FriendDto 확장해서 구현해야할 듯, Friend 관련 컨트롤러도 분리가 필요해보임
-//    // 친구 요청 응답
-//    @PostMapping("/friends/respond")
-//    public ResponseEntity<String> respondToFriendRequest(@RequestBody Long requestId, @RequestParam Boolean accept) {
-//        return ResponseEntity.ok(communityService.respondToFriendRequest(requestId, accept));
-//    }
 
 
-    // 커뮤니티 피드 조회
-    @GetMapping("/community/feed")
-    public ResponseEntity<List<CommunityPostDto>> getCommunityFeed() {
-        // 임의의 커뮤니티 피드 데이터 생성
-        List<CommunityPostDto> communityFeed = Arrays.asList(
-                CommunityPostDto.builder()
-                        .postId(1L)
-                        .authorId(100L)
-                        .postContent("첫 번째 커뮤니티 포스트 내용")
-                        .postTitle("첫 번째 커뮤니티 포스트")
-                        .creationDate("2023-11-10")
-                        .lastEdited("2023-11-10")
-                        .likeCount(15)
-                        .commentCount(4)
-                        .build(),
-                CommunityPostDto.builder()
-                        .postId(2L)
-                        .authorId(101L)
-                        .postContent("두 번째 커뮤니티 포스트 내용")
-                        .postTitle("두 번째 커뮤니티 포스트")
-                        .creationDate("2023-11-11")
-                        .lastEdited("2023-11-11")
-                        .likeCount(10)
-                        .commentCount(2)
-                        .build()
-        );
 
-        return ResponseEntity.ok(communityFeed);
-    }
-
-
-    // 커뮤니티 포스트 작성
-    @PostMapping("/community/posts")
-    public ResponseEntity<String> createCommunityPost(@RequestBody CommunityPostDto communityPostDto) {
-        CommunityPostDto newPost = CommunityPostDto.builder()
-                .postId(1L) // 임의의 게시물 ID
-                .authorId(communityPostDto.getAuthorId()) // 작성자 ID
-                .postContent(communityPostDto.getPostContent()) // 게시물 내용
-                .postTitle(communityPostDto.getPostTitle()) // 게시물 제목
-                .creationDate("2023-11-15") // 게시 날짜
-                .lastEdited("2023-11-15") // 마지막 수정 날짜
-                .likeCount(0) // 좋아요 수
-                .commentCount(0) // 댓글 수
-                .build();
-
-        String responseMessage = String.format("커뮤니티 포스트 생성 성공: %s", newPost.getPostTitle());
-        return ResponseEntity.ok(responseMessage);
-    }
-
-
-    // 커뮤니티 포스트 목록 조회
-    @GetMapping("/community/posts")
-    public ResponseEntity<List<CommunityPostDto>> getCommunityPosts() {
-        List<CommunityPostDto> posts = Arrays.asList(
-                CommunityPostDto.builder()
-                        .postId(1L)
-                        .authorId(101L)
-                        .postContent("첫 번째 포스트 내용")
-                        .postTitle("첫 번째 포스트")
-                        .creationDate("2023-11-15")
-                        .lastEdited("2023-11-15")
-                        .likeCount(10)
-                        .commentCount(5)
-                        .build(),
-                CommunityPostDto.builder()
-                        .postId(2L)
-                        .authorId(102L)
-                        .postContent("두 번째 포스트 내용")
-                        .postTitle("두 번째 포스트")
-                        .creationDate("2023-11-14")
-                        .lastEdited("2023-11-14")
-                        .likeCount(7)
-                        .commentCount(3)
-                        .build()
-        );
-
-        return ResponseEntity.ok(posts);
-    }
-
-
-    // 포스트 좋아요
-    @PostMapping("/community/posts/{postId}/like")
-    public ResponseEntity<String> likeCommunityPost(@PathVariable Long postId) {
-        CommunityPostDto likedPost = CommunityPostDto.builder()
-                .postId(postId)
-                .authorId(101L) // 임의의 작성자 ID
-                .postContent("이 포스트는 좋아요를 받았습니다.")
-                .postTitle("좋아요 받은 포스트")
-                .creationDate("2023-11-15")
-                .lastEdited("2023-11-15")
-                .likeCount(11) // 좋아요 수 증가
-                .commentCount(5)
-                .build();
-
-        String responseMessage = String.format("포스트 ID %d에 좋아요 성공", postId);
-        return ResponseEntity.ok(responseMessage);
-    }
 }
